refactor: Improve the content section and code reusability
